Реализован класс проверки правильности написания (отсутствия невозможных сочетаний) в тексте SpellingTextVerification
